+++ src/main/kotlin/dev/ekvedaras/hyperfquery/services/HyperfQuerySettingsConfigurable.kt
@@ -23,6 +23,7 @@ class HyperfQuerySettingsConfigurable(val project: Project) : SearchableConfigur
         settings.filterDataSources = settingsForm?.shouldFilterDataSources() ?: false
         settings.filteredDataSources = settingsForm?.filteredDataSources() ?: setOf()
         settings.tablePrefix = settingsForm?.tablePrefix()?.trim() ?: ""
+        settings.touch()
     }
 
     override fun reset() {